chore: extract SearchInput into component
