+++ src/WebComponent.scss
@@ -35,6 +35,7 @@
 @import "./components/Menu/MenuItemHeader.scss";
 @import "./components/Search/Search.scss";
 @import "./components/Search/SearchToggle.scss";
+@import "./components/Search/SearchInput.scss";
 @import "./components/Share/Share.scss";
 @import "./components/TopicMenu/TopicMenu.scss";
 @import "./components/TopicsMenu/TopicsMenu.scss";
