+++ src/components/Search/Search.scss
@@ -1,5 +1,5 @@
 /* stylelint-disable selector-class-pattern */
-@import '../../globals.scss';
+@import "../../globals.scss";
 
 .wkp-search {
   position: absolute;
@@ -7,135 +7,4 @@
   z-index: 2; /* To be above header div, in order to be clickable. */
   right: 480px;
   left: 400px;
-
-  .wkp-search-input {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: white;
-    border: 2px solid #b7b7b7;
-    height: 42px;
-  }
-
-  .wkp-search-button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: none;
-    color: white;
-    height: 42px;
-    width: 46px;
-    position: absolute;
-    top: 4px;
-    right: 52px;
-    font-size: 16px;
-
-    &.wkp-search-button-clear {
-      background: none;
-      color: #666;
-      right: 98px;
-    }
-
-    &.wkp-search-button-submit {
-      background-color: $brand-primary;
-    }
-  }
-
-  .react-autosuggest__input {
-    border: 0;
-    font-size: 14px;
-    padding-left: 15px;
-    border-radius: 0;
-    appearance: none;
-    flex: 2 1 auto;
-    height: 90%;
-
-    &::-ms-clear {
-      display: none;
-    }
-  }
-
-  .react-autosuggest__suggestions-container--open {
-    box-sizing: border-box;
-    background: white;
-    border: 2px solid #666;
-    margin-top: 5px;
-    width: 100%;
-    max-height: calc(100vh - 150px);
-    overflow: auto;
-  }
-
-  .react-autosuggest__section-title {
-    span {
-      color: #666;
-    }
-
-    .wkp-search-section-header {
-      & > span:first-child {
-        color: black;
-      }
-    }
-
-    .wkp-search-section-opener {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      background-color: #f5f5f5;
-      padding: 10px 15px;
-    }
-
-    :hover {
-      svg,
-      .MuiTypography-subtitle1 {
-        color: $brand-primary-hover;
-      }
-    }
-  }
-
-  .react-autosuggest__suggestions-list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-
-    li:first-child {
-      padding-top: 5px;
-    }
-
-    li:last-child {
-      padding-bottom: 5px;
-    }
-  }
-
-  .react-autosuggest__suggestion {
-    padding: 0 15px;
-    cursor: pointer;
-
-    .wkp-search-suggestion,
-    .wkp-search-suggestion-subtitled {
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-    }
-
-    .wkp-search-suggestion {
-      min-height: 25px;
-      padding: 3px 0;
-    }
-
-    .wkp-search-suggestion-subtitled {
-      min-height: 50px;
-    }
-
-    &.react-autosuggest__suggestion--highlighted {
-      p {
-        color: $brand-primary-hover !important;
-      }
-    }
-
-    &:hover {
-      p {
-        color: $brand-primary-hover !important;
-      }
-    }
-  }
 }
