+++ src/menus/StsMenu/StsMenu.js
@@ -18,7 +18,7 @@ import { DV_KEY } from "../../utils/constants";
 import useHasScreenSize from "../../utils/useHasScreenSize";
 import useHighlightLayer from "../../utils/useHighlightLayer";
 import DvFeatureInfoTitle from "../../config/ch.sbb.direktverbindungen/DvFeatureInfoTitle/DvFeatureInfoTitle";
-import { SearchInput } from "../../components/Search/Search";
+import SearchInput from "../../components/Search/SearchInput";
 
 const boxShadow =
   "0px 5px 5px -3px rgb(0 0 0 / 20%), 0px 8px 10px 1px rgb(0 0 0 / 14%), 0px 3px 14px 2px rgb(0 0 0 / 12%)";
